Big Update
Significant structural changes to my code. I cleaned up the code in
general and tried to shorten it where I could, as well as try to
unclutter the code. I added the "double-down" functionality, and made
the program automatically handles Aces in the user's hand. Next updates
will try to further shorten the code, mainly the `turn` function.